Extract shared DFC context normalisation into dfcContext.js
Addresses Copilot review feedback: deduplicate DFC_CONTEXT_W3ID,
DFC_CONTEXT_PRELOADED, and normalizeContext() into a single shared
module imported by all three call sites. Fixes double-quote lint
error in index.js.
